Scorecard/dot_inspections_etl/requirements.txt
pyodbc>=4.0.30
python-dotenv>=0.19.0
xmltodict>=0.13.0